Terraform and refactorings
