+++ source/B2B.Transactions.CimMessageAdapter/Schema/SchemaProvider.cs
@@ -12,13 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Linq;
 using System.Threading.Tasks;
 using System.Xml;
 using System.Xml.Schema;
-using B2B.CimMessageAdapter.Schema;
 
-namespace B2B.CimMessageAdapter
+namespace B2B.CimMessageAdapter.Schema
 {
     public class SchemaProvider : ISchemaProvider
     {
